moved all components to the dedicated folder
