+++ client/src/components/Footer.js
@@ -1,11 +1,7 @@
-
-
 import React from "react";
-import Nav from "../nav";
+import Nav from "./Nav";
 import { FaInstagramSquare } from "react-icons/fa";
 import { FaFacebookF } from "react-icons/fa";
-
-
 
 export default function Footer() {
   return (
@@ -20,7 +16,6 @@ export default function Footer() {
             <span><FaInstagramSquare /></span>
             <span><FaFacebookF /></span>
         </div>
-        {/* <Button /> */}
     </div>
   );
 }
