+++ client/src/components/Nav.js
@@ -1,9 +1,8 @@
 import React from "react";
 
-
-export default function Nav(props) {
+export default function Nav() {
   return (
-    <nav style={{ color: props.color, background:props.backgroundColor, padding: props.padding }}>
+    <nav>
         <li><a href="https://www.pratibhanakshatra.org/public/storage/setting/1594904931pratibha.jpg">Home</a></li>
         <li><a href="https://www.pratibhanakshatra.org/public/storage/setting/1594904931pratibha.jpg">About</a></li>
         <li><a href="https://www.pratibhanakshatra.org/public/storage/setting/1594904931pratibha.jpg">Blogs</a></li>
