+++ client/src/components/Video.js
@@ -1,11 +1,11 @@
 import React from "react";
-import Skyscraper from "./images/PlaceholderImage.jpeg";
+import Skyscraper from "../images/PlaceholderImage.jpeg";
 
 const Video = (props) => {
   console.log(props);
   return (
     <div className="videoWithOverlay">
-      <img className="video" src={Skyscraper} />
+      <img className="video" src={Skyscraper} alt="" />
          {/* autoPlay loop muted /> */}
       <div className="textOverlay">
         <p>Placeholder Text for Video Component Overlay</p>
